module_products/serializers.py
from rest_framework import serializers
from models.product import Product

class ProductSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Product
        fields = ('id', 'title','price','description')